Desktop/7/server_app/requirement.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==2.8
DBUtils==1.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-cors-headers==3.2.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.2.0
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jwcrypto==0.6.0
MarkupPy==1.14
MarkupSafe==1.1.1
Naked==0.1.31
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
pycparser==2.19
pycryptodome==3.9.6
PyJWT==1.7.1
PyMySQL==0.9.3
python-jwt==3.2.5
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
uritemplate==3.0.1
urllib3==1.25.7
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
